* commit '0ace69886a24d39dcaa6cf1afa1982d4644f16c8':
  ExposureKey expiration is now depending on upload time
  Add exposed keys upload functionality.
  Add dependency injection extensions. TODO: integration
  Change cosmot -> sql Change profilr id to long

+++ Sygic.Corona.Application/Commands/SendPushNotificationCommand.cs
@@ -4,10 +4,10 @@ namespace Sygic.Corona.Application.Commands
 {
     public class SendPushNotificationCommand : IRequest
     {
-        public uint ProfileId { get; }
+        public long ProfileId { get; }
         public object Message { get; }
 
-        public SendPushNotificationCommand(uint profileId, object message)
+        public SendPushNotificationCommand(long profileId, object message)
         {
             ProfileId = profileId;
             Message = message;
