+++ Sygic.Corona.Application/Queries/GetContactsQuery.cs
@@ -6,9 +6,9 @@ namespace Sygic.Corona.Application.Queries
     public class GetContactsQuery : IRequest<GetContactsResponse>
     {
         public string DeviceId { get; }
-        public uint ProfileId { get; }
+        public long ProfileId { get; }
 
-        public GetContactsQuery(string deviceId, uint profileId)
+        public GetContactsQuery(string deviceId, long profileId)
         {
             DeviceId = deviceId;
             ProfileId = profileId;
